Moved btcutil/koblitz to crypto/koblitz.
