+++ btcutil/address.go
@@ -10,7 +10,7 @@ import (
 	"errors"
 	"github.com/mit-dci/lit/btcutil/base58"
 	"github.com/mit-dci/lit/btcutil/chaincfg"
-	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/crypto/ripemd160"
 )
