+++ btcutil/hdkeychain/extendedkey.go
@@ -20,8 +20,8 @@ import (
 	"github.com/mit-dci/lit/btcutil/base58"
 	"github.com/mit-dci/lit/btcutil/chaincfg"
 	"github.com/mit-dci/lit/btcutil/chaincfg/chainhash"
-	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/coinparam"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"math/big"
 )
