+++ btcutil/txscript/opcode.go
@@ -12,7 +12,7 @@ import (
 	"fmt"
 	"hash"
 
-	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/btcutil/chaincfg/chainhash"
 	"github.com/mit-dci/lit/crypto/fastsha256"
 	"github.com/mit-dci/lit/crypto/ripemd160"
