+++ crypto/koblitz/README.md
@@ -1,0 +1,4 @@
+# lit/crypto/koblitz
+
+This used to be in btcutil but was moved here because it was a more
+logical place for it than where it was.
