+++ crypto/koblitz/precompute.go
