+++ crypto/koblitz/privkey.go
