+++ litrpc/lndcrpcclient.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"github.com/mit-dci/lit/btcutil/hdkeychain"
-	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/coinparam"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/lndc"
 	"github.com/mit-dci/lit/lnutil"
 	"github.com/mit-dci/lit/portxo"
