+++ lndc/conn.go
@@ -8,7 +8,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/lnutil"
 	"github.com/mit-dci/lit/logging"
 )
