+++ lnp2p/peermgr.go
@@ -4,7 +4,7 @@ package lnp2p
 import (
 	"crypto/ecdsa"
 	"fmt"
-	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/btcutil/hdkeychain"
 	"github.com/mit-dci/lit/eventbus"
 	"github.com/mit-dci/lit/lndc"
