+++ lnutil/curvelib_test.go
@@ -3,7 +3,7 @@ package lnutil
 import (
 	"testing"
 
-	"github.com/mit-dci/lit/btcutil/koblitz"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/btcutil/chaincfg/chainhash"
 )
