+++ qln/fund.go
@@ -2,8 +2,8 @@ package qln
 
 import (
 	"fmt"
-	"github.com/mit-dci/lit/btcutil/koblitz"
 	"github.com/mit-dci/lit/consts"
+	"github.com/mit-dci/lit/crypto/koblitz"
 	"github.com/mit-dci/lit/elkrem"
 	"github.com/mit-dci/lit/lnutil"
 	"github.com/mit-dci/lit/logging"
